Reorganize scripts (#14)
* Reorganize scripts, add several README
