+++ scripts/analysis/README.md
@@ -1,0 +1,3 @@
+### Description
+
+This module contains scripts for the analysis of different data extracted from Kotlin projects.
