+++ scripts/analysis/clones/column_names_utils.py
