+++ scripts/data_collection/filter_dataset.py
